requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0
